+++ database/migrations/2021_10_16_150234_create_hints_table.php
@@ -19,7 +19,7 @@ class CreateHintsTable extends Migration
             $table->text('description_ar');
             $table->text('description_en');
 
-            $table->integer('media_id')->unsigned();
+            $table->bigInteger('media_id')->unsigned();
             $table->foreign('media_id')->references('id')->on('media');
 
             $table->timestamps();
