+++ database/migrations/2021_10_16_150235_create_questions_table.php
@@ -19,13 +19,13 @@ class CreateQuestionsTable extends Migration
             $table->text('title_ar');
             $table->text('title_en');
 
-            $table->integer('media_id')->unsigned();
+            $table->bigInteger('media_id')->unsigned();
             $table->foreign('media_id')->references('id')->on('media');
 
             $table->string('answers');
             $table->string('right_answer');
 
-            $table->integer('hint_id')->unsigned();
+            $table->bigInteger('hint_id')->unsigned();
             $table->foreign('hint_id')->references('id')->on('hints');
 
             $table->timestamps();
