+++ database/migrations/2021_10_16_150246_create_exams_table.php
@@ -21,7 +21,7 @@ class CreateExamsTable extends Migration
             $table->string('name_ar');
             $table->string('name_en');
 
-            $table->integer('media_id')->unsigned();
+            $table->bigInteger('media_id')->unsigned();
             $table->foreign('media_id')->references('id')->on('media');
 
             $table->timestamps();
