+++ database/migrations/2021_10_26_140304_create_user_subjects_table.php
@@ -16,10 +16,10 @@ class CreateUserSubjectsTable extends Migration
         Schema::create('user_subjects', function (Blueprint $table) {
             $table->id();
 
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users');
 
-            $table->integer('subject_id')->unsigned();
+            $table->bigInteger('subject_id')->unsigned();
             $table->foreign('subject_id')->references('id')->on('subjects');
 
             $table->timestamps();
